requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14      
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0     
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
poppler-utils==0.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
qrdet==2.5
qreader==3.14
quadrilateral-fitter==1.12
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2